lambda/tests/requirements.txt
boto3
humanfriendly
jmespath
jsonpath
moto
pytest
pyyaml
voluptuous
